Refactor device configuration into reusable audio.devices module
- Create new raspberry_pi/audio/ package directory
- Extract USB_ADAPTER, Statue enum, dynConfig, and configure_devices() to devices.py
- Update tone_detect_test.py to import from the new module
- Move tone frequency assignment to after device configuration
- Remove unused json import
- Maintain all existing functionality with better code organization

This refactoring allows device configuration to be reused by other scripts
in the project, promoting code reuse and maintainability.
