+++ raspberry_pi/audio/__init__.py
@@ -1,0 +1,1 @@
+# Audio module for Missing Link project
